requirements-standalone.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
bottle==0.13.4
click==8.3.0
fastapi==0.109.0
h11==0.16.0
httptools==0.7.1
idna==3.11
macholib==1.16.3
packaging==25.0
pillow==12.0.0
proxy_tools==0.1.0
pydantic==2.12.3
pydantic_core==2.41.4
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-Security==12.0
pyobjc-framework-WebKit==12.0
python-dotenv==1.2.1
python-multipart==0.0.9
pywebview==4.4.1
PyYAML==6.0.3
setuptools==80.9.0
sniffio==1.3.1
starlette==0.35.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
